requirements.txt
setuptools
wheel
spacy
matplotlib
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
ipykernel
google-cloud-bigquery
google-cloud-storage
pyarrow
db-dtypes
fastapi
uvicorn
transformers
